Centered heading on session alert modal
