--- src/domain/app/sessionPrompt/sessionPrompt.module.scss
@@ -1,20 +1,0 @@
-@import 'variables';
-@import 'mixins';
-
-.sessionExpired {
-  text-align: center;
-  p {
-    padding-bottom: $largeMargin;
-  }
-}
-
-.goBackButton {
-  @include button();
-  margin-bottom: $largeMargin;
-}
-
-.modal {
-  h1 {
-    text-align: center;
-  }
-}
